# gm-center/gmc_api_gateway/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Gedge GM-Center Swagger API.",
        "title": "Gedge GM-Center Swagger API",
        "contact": {
            "name": "GM-Center",
            "url": "https://gedge-platform.github.io/gm-center/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "101.79.1.138:8012",
    "basePath": "/gmcapi/v2",
    "paths": {
...
        "/cronjobs": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get Cronjob List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes"
                ],
                "summary": "Show List Cronjob",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the Workspace",
                        "name": "workspace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name of the Cluster",
                        "name": "cluster",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "name of the Project",
                        "name": "project",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CRONJOB"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create Cronjob",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes"
                ],
                "summary": "Create Cronjob",
                "parameters": [
                    {
                        "description": "Cronjob Info Body",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "name of the Cluster",
                        "name": "cluster",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Workspace",
                        "name": "workspace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Project",
                        "name": "project",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/cronjobs/{name}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "get Cronjob Details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes"
                ],
                "summary": "Show detail Cronjob",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the Cronjob",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Workspace",
                        "name": "workspace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Cluster",
                        "name": "cluster",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Project",
                        "name": "project",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CRONJOB"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete Cronjob",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes"
                ],
                "summary": "Delete Cronjob",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name of the Cronjob",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Workspace",
                        "name": "workspace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Cluster",
                        "name": "cluster",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name of the Project",
                        "name": "project",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Active": {
            "type": "object",
            "properties": {
                "kind": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
...
        "model.CRONJOB": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Active"
                    }
                },
                "annotations": {},
                "cluster": {
                    "type": "string"
                },
                "concurrencyPolicy": {
                    "type": "string"
                },
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Containers"
                    }
                },
                "creationTimestamp": {
                    "type": "string"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EVENT"
                    }
                },
                "failedJobsHistoryLimit": {
                    "type": "integer"
                },
                "label": {},
                "lastScheduleTime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "project": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                },
                "successfulJobsHistoryLimit": {
                    "type": "integer"
                },
                "user": {
                    "type": "string"
                },
                "workspace": {
                    "type": "string"
                }
            }
        },
...
        "model.Containers": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.EVENT": {
            "type": "object",
            "properties": {
                "cluster": {
                    "type": "string"
                },
                "eventTime": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "\"Type \\\"Bearer \\\" and then your API Token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}